% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{clip_raster_from_shapefile}
\alias{clip_raster_from_shapefile}
\title{A simple function to clip a raster using a shapefile}
\usage{
clip_raster_from_shapefile(raster, shapefile)
}
\arguments{
\item{shapefile}{A shapefile object.}

\item{raset}{A raster object.}
}
\value{
The raster file with NAs outside of the shapefile mask.
}
\description{
A simple function to clip a raster using a shapefile
}
