% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_france_raster_template}
\alias{get_france_raster_template}
\title{Create a raster to serve as template in other functions}
\usage{
get_france_raster_template(res_km)
}
\arguments{
\item{res_km}{A numeric (int or float) that represents the average spatial resolution of the grid (each cell will be approx. \code{res}*\code{res} km)}
}
\value{
A raster with no values assigned.
}
\description{
The raster created is centered on France It is used by several
other functions to create identical raster grids.
}
