% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_france_shapefile}
\alias{get_france_shapefile}
\title{Download a shapefile of France}
\usage{
get_france_shapefile(borders = "national")
}
\arguments{
\item{borders}{A string. Should be \code{"national"} (default) for outside borders or \code{"dpt"} for french 'département' borders.}
}
\value{
A shapefile of france.
}
\description{
Download a shapefile of France
}
