% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_meteofr_urls}
\alias{get_meteofr_urls}
\title{Fetch and filter urls of meteo-france datasets available on data.gouv.fr}
\usage{
get_meteofr_urls()
}
\value{
A vector of strings, which are links to CSVs of meteo-france
available online.
}
\description{
Gets the urls from the API and filters out "documentation" urls and
departement 99 (whose definition is unclear). Only the years 1950-2024 are
kept.
}
