% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{ggplot_categorical_df_on_france_map}
\alias{ggplot_categorical_df_on_france_map}
\title{Create a ggplot that shows scattered locations on a map with their categories}
\usage{
ggplot_categorical_df_on_france_map(
  base_map,
  df,
  LON = "x",
  LAT = "y",
  column = NULL
)
}
\arguments{
\item{base_map}{A base ggplot on which to draw the grid (usually, obtained from \code{\link{ggplot_get_france_base_map}})}

\item{df}{A data.frame (or a CSV path to fetch) with columns LON and LAT corresponding to coordinates in CRS:4326}

\item{LON}{A string, the name of a column in df (Longitude coordinate)}

\item{LAT}{A string, the name of a column in df (Latitude coordinate)}

\item{column}{A string, name of the column where the observations are stored (values must be binary c(0, 1)).}
}
\value{
A ggplot object
}
\description{
Create a ggplot that shows scattered locations on a map with their categories
}
\seealso{
\code{\link{ggplot_get_france_base_map}}
}
