% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{ggplot_categorical_raster_on_france_map}
\alias{ggplot_categorical_raster_on_france_map}
\title{Create a ggplot that shows a grid of categorical values on the map of France}
\usage{
ggplot_categorical_raster_on_france_map(base_map, raster, layer_name)
}
\arguments{
\item{base_map}{A base ggplot on which to draw the grid (usually, obtained from \code{\link{ggplot_get_france_base_map}})}

\item{raster}{Either a string (path to a raster .tif file) or a SpatRaster object.}

\item{layer_name}{String that corresponds to the name of the layer to plot.}
}
\value{
A ggplot object
}
\description{
Create a ggplot that shows a grid of categorical values on the map of France
}
\seealso{
\code{\link{ggplot_get_france_base_map}}
}
