% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{ggplot_get_france_base_map}
\alias{ggplot_get_france_base_map}
\title{Create a ggplot that shows a basic map of France}
\usage{
ggplot_get_france_base_map(borders_type = "national")
}
\arguments{
\item{borders_type}{A string. Direct call to \code{\link{get_france_shapefile}}. Should be \code{"national"} (default) for outside borders or \code{"dpt"} for french 'département' borders.}
}
\value{
A ggplot object
}
\description{
Create a ggplot that shows a basic map of France
}
\seealso{
\code{\link{get_france_shapefile}}
}
