% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{ggplot_points_scattered_on_france_map}
\alias{ggplot_points_scattered_on_france_map}
\title{Create a ggplot that shows weather stations locations}
\usage{
ggplot_points_scattered_on_france_map(base_map, csv_path)
}
\arguments{
\item{base_map}{A base ggplot on which to draw the stations locations (usually, obtained from \code{\link{ggplot_get_france_base_map}})}

\item{csv_path}{A string that is the path to a CSV with LON and LAT columns (CRS 4326).}
}
\value{
A ggplot object
}
\description{
Creates a ggplot that shows weather stations locations
on a basic map of France
}
\seealso{
\code{\link{ggplot_get_france_base_map}}
}
