% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{ggplot_quantitative_raster_on_france_map}
\alias{ggplot_quantitative_raster_on_france_map}
\title{Create a ggplot that shows a grid of continuous values on the map of France}
\usage{
ggplot_quantitative_raster_on_france_map(
  base_map,
  raster_path,
  colname,
  unit = "°C",
  limits = NULL
)
}
\arguments{
\item{base_map}{A base ggplot on which to draw the grid (usually, obtained from \code{\link{ggplot_get_france_base_map}})}

\item{raster_path}{A string that is the path to a raster (.tif) file.}

\item{colname}{Name of the column for which the values will be plotted.}

\item{unit}{Name of the unit for colname.}

\item{limits}{Value limits for the color bars. Can either be NULL (auto limits by steps of 5) or a list of 2 values (c(min, max)).}
}
\value{
A ggplot object
}
\description{
Create a ggplot that shows a grid of continuous values on the map of France
}
\seealso{
\code{\link{ggplot_get_france_base_map}}
}
