% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{group_CLC_categories}
\alias{group_CLC_categories}
\title{Group together the selected categories of a raster}
\usage{
group_CLC_categories(
  raster,
  cats_2_group,
  new_group_name = "unmodelled",
  new_group_color = "#000000"
)
}
\arguments{
\item{raster}{A basic SpatRaster object. Must have values & land_cover fields}

\item{cats_2_group}{A list of categories to group together (must be a sample of \code{terra::cats(raster)[[1]]$land_cover})}

\item{new_group_name}{A string. Self-explanatory.}

\item{new_group_color}{A string. Self-explanatory. See \code{help(col2rgb)} for accepted formats.}
}
\value{
The raster with its new categories.
}
\description{
Reduces the number of categories in a raster by grouping together the
categories specified by the user. Conserves original colors.
}
