% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{import_and_transform_corine_raster}
\alias{import_and_transform_corine_raster}
\title{Load weather dataframe and pre-proccess it to a grid}
\usage{
import_and_transform_corine_raster(
  corine_tiff_path,
  res_km = 10,
  verbose = TRUE
)
}
\arguments{
\item{res_km}{A numeric (int or float) that represents the average spatial resolution of the grid (each cell will be approx. \code{res}*\code{res} km)}

\item{data_path}{A string that is the path to a CSV (saved from \code{\link{select_year_round_obs}} output).}

\item{colname}{A string corresponding to a columns in the CSV imported. Value that will be interpolated.}
}
\value{
A raster of values covering france
}
\description{
Loads dataframe saved as CSV from \code{\link{select_year_round_obs}} outputs
and then either returns it as a raster grid covering France.
}
\seealso{
\code{\link{interpolate_meshgrid}}, \code{\link{clip_raster_from_shapefile}}, \code{\link{select_year_round_obs}}
}
