% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{import_biodicapt_land_surveys}
\alias{import_biodicapt_land_surveys}
\title{Load and format BIODICAPT surveys obtained from networks}
\usage{
import_biodicapt_land_surveys(xlsx_paths)
}
\arguments{
\item{xlsx_paths}{A vector of 12 strings that are paths to raster files.}
}
\value{
A dataframe (a concatenation of all surveys)
}
\description{
Load and format BIODICAPT surveys obtained from networks
}
