% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{interpolate_scattered_data}
\alias{interpolate_scattered_data}
\title{Create a raster from scattered data}
\usage{
interpolate_scattered_data(df, val = "TM", res_km = 10, linear = TRUE)
}
\arguments{
\item{df}{A dataframe with columns c("LON", "LAT", val).}

\item{val}{A string, which is the name for the value in \code{df} that will be interpolated.}

\item{res_km}{A numeric (int or float) that represents the average spatial resolution of the grid (each cell will be approx. \code{res}*\code{res} km)}

\item{linear}{A boolean, \code{TRUE} (default) for linear interpolation, \code{FALSE} for spline interpolation.}
}
\value{
A raster grid in the WGS84 coordinate system.
}
\description{
Inspired by meshgrid in Python. Takes a dataframe of geographically scattered
points and computes a grid that interpolates values between each point, using
linear or spline method.
}
\seealso{
\code{\link{load_url_2_df}}, \code{\link{select_year_round_obs}}
}
