% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{LEGACY_ggplot_CSV_points_scattered_on_france_map}
\alias{LEGACY_ggplot_CSV_points_scattered_on_france_map}
\title{Create a ggplot that shows the locations of CSV rows}
\usage{
LEGACY_ggplot_CSV_points_scattered_on_france_map(base_map, csv_path, category)
}
\arguments{
\item{base_map}{A base ggplot on which to draw the stations locations (usually, obtained from \code{\link{ggplot_get_france_base_map}})}

\item{csv_path}{A string that is the path to a CSV with LON and LAT columns (CRS 4326).}

\item{category}{A string that is on of the columns in the imported CSV. Allows the user to control color and fill attributes of the points plotted on the ggplot after its generation. (Default is NULL -> use of default colors).}
}
\value{
A ggplot object
}
\description{
Create a ggplot that shows the locations of CSV rows
}
\seealso{
\code{\link{ggplot_get_france_base_map}}
}
