% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{LEGACY_ggplot_xy_points_scattered_on_france_map}
\alias{LEGACY_ggplot_xy_points_scattered_on_france_map}
\title{Create a ggplot that shows weather stations locations}
\usage{
LEGACY_ggplot_xy_points_scattered_on_france_map(
  base_map,
  df,
  LON = "x",
  LAT = "y"
)
}
\arguments{
\item{base_map}{A base ggplot on which to draw the stations locations (usually, obtained from \code{\link{ggplot_get_france_base_map}})}

\item{df}{A data.frame with columns LON and LAT corresponding to coordinates in CRS:4326}

\item{LON}{A string, the name of a column in df (Longitude coordinate)}

\item{LAT}{A string, the name of a column in df (Latitude coordinate)}
}
\value{
A ggplot object
}
\description{
Create a ggplot that shows weather stations locations
}
\seealso{
\code{\link{ggplot_get_france_base_map}}
}
