% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{load_url_2_df}
\alias{load_url_2_df}
\title{Import and concatenate meteo france dataset to a data.frame}
\usage{
load_url_2_df(
  url_list,
  keep_year,
  keep_columns = c("NUM_POSTE", "NOM_USUEL", "LAT", "LON", "AAAAMM", "TX", "TN", "TM"),
  verbose = TRUE
)
}
\arguments{
\item{url_list}{A vector of strings containing CSVs urls.}

\item{keep_year}{A string corresponding to a year in 1950-2024 (fmt: "YYYY").}

\item{keep_columns}{A vector of strings corresponding to the names of columns that will be kept after loading the CSVs.}

\item{verbose}{A boolean. If \code{TRUE} activates progressbar and messages.}
}
\value{
A single dataframe containing selected meteo france data.
}
\description{
Takes a vector of url strings and keeps the data corresponding to the year
that is indicated to import the data into one dataframe.
}
\seealso{
\code{\link{get_meteofr_urls}}
}
