% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{monthly_2_yearly_rasters}
\alias{monthly_2_yearly_rasters}
\title{Load and format monthly collected rasters to a single annual raster}
\usage{
monthly_2_yearly_rasters(raster_paths, res_km, fun = mean)
}
\arguments{
\item{raster_paths}{A vector of 12 strings that are paths to raster files.}

\item{res_km}{A numeric (int or float) that represents the average spatial resolution of the grid (each cell will be approx. \code{res}*\code{res} km)}

\item{fun}{A function for the aggregation of month data to annual data (usually one of: mean, median, sum)}
}
\value{
A raster of values covering france
}
\description{
Loads 12 raster (each one corresponding to a different month) from given
paths. Then, crop to France extent and reprojects to a new template grid.
}
\seealso{
\code{\link{interpolate_meshgrid}}, \code{\link{clip_raster_from_shapefile}}, \code{\link{select_year_round_obs}}
}
