% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{my_custom_ggplot_theme}
\alias{my_custom_ggplot_theme}
\title{Create a custom ggplot theme that can be applied to any ggplot figure}
\usage{
my_custom_ggplot_theme(figure, with_palette = TRUE)
}
\arguments{
\item{figure}{A ggplot2 figure.}

\item{with_palette}{A boolean. If TRUE (default), adds custom color, fill, shape and size scales.}
}
\value{
A ggplot figure.
}
\description{
Create a custom ggplot theme that can be applied to any ggplot figure
}
