% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{normal_density_plot_with_thresholds}
\alias{normal_density_plot_with_thresholds}
\title{Create a density ggplot of a normal function}
\usage{
normal_density_plot_with_thresholds(
  mean_val = 1,
  sd_val = 1.5,
  thresholds = NULL
)
}
\arguments{
\item{mean_val}{Mean value of the normal law.}

\item{sd_val}{Standard deviation of the normal law.}

\item{thresholds}{A vector of numeric values (optional), will add vertical lines corresponding to those on the plot.}
}
\description{
Does what the title says but also can show vertical lines for custom thresholds.
Also, show shaded grey areas for values outside the 95\% CI.
}
