% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{plot_grid_scattered_densities}
\alias{plot_grid_scattered_densities}
\title{Create a grid of density and/or scatter points}
\usage{
plot_grid_scattered_densities(
  x,
  y,
  category,
  df_points = NULL,
  df_lines = NULL
)
}
\arguments{
\item{x}{A string, the name of a column in both df (x-axis).}

\item{y}{A string, the name of a column in both df (y-axis).}

\item{category}{A string, the name of a column in both df, that has qualitative values.}

\item{df_points}{A data.frame with columns x, y and z.}

\item{df_lines}{A data.frame with columns x, y and z.}
}
\description{
Create a grid of density and/or scatter points
}
