% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{pretty_range}
\alias{pretty_range}
\title{Compute a range and snap limits to closest multiples}
\usage{
pretty_range(x, multiple = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{multiple}{A numeric. The multiple to which the limits will snap to.}
}
\value{
A vector of limits c(min, max)
}
\description{
Compute a range and snap limits to closest multiples
}
