% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_figures.R
\name{save_figure_harmonized}
\alias{save_figure_harmonized}
\title{Save a ggplot and its legend separately}
\usage{
save_figure_harmonized(ggplot_obj, save_path, rescale_legend = c(4, 4))
}
\arguments{
\item{ggplot_obj}{A ggplot object to be saved as PDF}

\item{save_path}{The path were the file should be saved. Must end with ".pdf". The legend will be saved with the same filename, appended with '_legend'.}

\item{rescale_legend}{A vector of 2 numeric values. Change if the legend is too wide/tall to fit inside the default square. Default is c(1,1) (no rescaling). Increase first value for width, second value for height.}
}
\description{
Create to separate plots from one ggplot entry: one for the plot and one
for the legend, then saves them as separate files. If there is no legend,
only one file is saved for the figure.
}
