% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{select_year_round_obs}
\alias{select_year_round_obs}
\title{Turn monthly data from meteo-france to annual data}
\usage{
select_year_round_obs(df, mode = "average")
}
\arguments{
\item{df}{A dataframe (from \code{\link{load_url_2_df}}).}

\item{mode}{A string for the grouping method. Should be either \code{"average"} (default) or \code{"median"}.}
}
\value{
A single dataframe containing annual data.
}
\description{
Groups the dataframe of monthly collected temperature by meteo france into
annual data. Can choose between "average" or "median" grouping.
Weather stations with less than 12 data points (i.e. 12 month of collected
data in the year) are discarded.
}
\seealso{
\code{\link{load_url_2_df}}
}
