% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{simplify_CLC}
\alias{simplify_CLC}
\title{Group together similar categories of CORINE Land Cover}
\usage{
simplify_CLC(
  clc_raster,
  group_urban = TRUE,
  group_crops = FALSE,
  group_forests = TRUE,
  group_wetlands = TRUE,
  group_water = TRUE
)
}
\arguments{
\item{clc_raster}{The CLC raster to modify.}

\item{group_urban}{A boolean that controls the grouping (TRUE, default) of urban areas into a single category.}

\item{group_crops}{A boolean that controls the grouping (TRUE) of agricultural areas into a single category. Default is FALSE.}

\item{group_forests}{A boolean that controls the grouping (TRUE, default) of forests and semi-natural areas into a single category.}

\item{group_wetlands}{A boolean that controls the grouping (TRUE, default) of wetlands into a single category.}

\item{group_water}{A boolean that controls the grouping (TRUE, default) of water bodies into a single category.}
}
\value{
The CLC raster with its new categories (layer is labelled: LABEL1).
}
\description{
The CORINE Land Cover (CLC) raster has many sub-categories that may or may not be
useful to the user depending on usage. This function enables to group the
subcategories into their respective categories to minimize the number of
different variables. The function replaces layer names and color table.
}
